import lib clear

+++ src/tw/edu/ttu/pre_towerdefender/GameScreenActivity.java
@@ -1,6 +1,5 @@
 package tw.edu.ttu.pre_towerdefender;
 
-import gameObject.tower.Soldier;
 import java.io.IOException;
 import java.util.Random;
 import gameSystem.gameObjectSystem.IDType;
@@ -8,7 +7,6 @@ import gameSystem.gameObjectSystem.ObjectHandler;
 import java.util.ArrayList;
 import java.util.Timer;
 import java.util.TimerTask;
-import tw.edu.ttu.pre_towerdefender.R.drawable;
 import gameviews.constants.Constant;
 import gameSystem.gameObjectSystem.gameObjectInfo.ObjectInfoReader;
 import com.metaio.sdk.ARViewActivity;
@@ -18,8 +16,6 @@ import com.metaio.sdk.jni.IMetaioSDKCallback;
 import com.metaio.sdk.jni.Rotation;
 import com.metaio.sdk.jni.TrackingValues;
 import com.metaio.sdk.jni.TrackingValuesVector;
-import com.metaio.sdk.jni.Vector2d;
-import com.metaio.sdk.jni.Vector2di;
 import com.metaio.sdk.jni.Vector3d;
 import com.metaio.tools.SystemInfo;
 import com.metaio.tools.io.AssetsManager;
@@ -27,7 +23,6 @@ import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Paint;
-import android.graphics.PorterDuff.Mode;
 import android.graphics.RectF;
 import android.graphics.drawable.Drawable;
 import android.hardware.Camera.CameraInfo;
@@ -40,7 +35,6 @@ import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.ViewGroup.LayoutParams;
 import android.widget.Button;
-import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.ProgressBar;
 import android.widget.TextView;
@@ -199,8 +193,8 @@ public class GameScreenActivity extends ARViewActivity {
 			params.width = (int)(Constant.imageSize[i][0] * Constant.wRatio);
 	        params.height =(int)( Constant.imageSize[i][1] * Constant.hRatio);
 	        imageArray.get(i).setLayoutParams(params);
-	        imageArray.get(i).setX((float)Constant.imageSize[i][2]*Constant.wRatio);
-	        imageArray.get(i).setY((float) (Constant.imageSize[i][3]*Constant.hRatio));
+	        imageArray.get(i).setX(Constant.imageSize[i][2]*Constant.wRatio);
+	        imageArray.get(i).setY(Constant.imageSize[i][3]*Constant.hRatio);
 	        
 		}
